modified paypal payment and payout, so that we store their fund

+++ paypal/models.py
@@ -12,4 +12,4 @@ class Transaction(models.Model):
 
 class UserFund(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL , on_delete=models.CASCADE)
-    fund = models.IntegerField()
+    fund = models.FloatField()
